requirements.txt
ffmpeg_python==0.2.0
protobuf==3.20.1
pydub==0.25.1
google-genai==1.13.0
python-dotenv==1.1.0